+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - English circled mode
 - English dingbat circled sans-serif mode
+- Tai Tham Hora numeral system
 ## [1.0] - 2025-10-20
 ### Added
 - Sora Sompeng numeral system